Add j-invariant to public interface

+++ skimage/restoration/j_invariant.py
@@ -7,7 +7,7 @@ from ..util import img_as_float
 
 def interpolate_image(image, multichannel=False):
     spatialdims = image.ndim if not multichannel else image.ndim - 1
-    conv_filter = ndi.generate_binary_structure(spatialdims, 1)
+    conv_filter = ndi.generate_binary_structure(spatialdims, 1).astype(float)
     conv_filter.ravel()[conv_filter.size // 2] = 0
     conv_filter /= conv_filter.sum()
